+++ src/math/sparmatblk.cc
@@ -14,7 +14,7 @@ SparMatBlk::SparMatBlk()
     alloc_   = 0;
     row_     = nullptr;
     blocks_  = nullptr;
-    colidx_  = new unsigned[2];
+    colidx_  = new size_t[2];
     colidx_[0] = 0;
 }
 
@@ -45,7 +45,7 @@ void SparMatBlk::allocate(size_t alc)
         alloc_ = alc;
         
         delete[] colidx_;
-        colidx_ = new unsigned[alc+1];
+        colidx_ = new size_t[alc+1];
         for ( size_t n = 0; n <= alc; ++n )
             colidx_[n] = n;
     }
